backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.121.3",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "weaviate-client>=4.18.1",
]
